Hack to get OS X + tmux copy/paste to work.
This is OS X specific, the default-command in tmux.conf must be
modified to be cross-platform. I'll do it the first time I pull
my dotfiles on a non OS X box. It looksl ike
reattach-to-user-namespace properly fails when not on OS X so it
should be easy to have something that's cross-platform.